offchain-content
> re-use feedRef type for offchain messages
a second type feels like it will introduce type switches on two things
which are mostly the same otherwise.

> unexport FeedRef ID field
might not be the best decision but usinig it straightup as
librarian.Addr turned into the problem we initially feared.

they get stored alright but when loading them back, how do you know the
type of it?

add PubKey() to FeedRef
add StoredAddr() to FeedRef

also:
* interop: use ssb-content-stream@3
* support offchain in contacts idx
* add offchain tests
* add tests for valid refs
* use different method for offchain feeds
* extract legacyDrain
* add private unbox